working merge of v0.8.5 and screenManager
