+++ core/src/com/bitsavior/screens/GameScreen.java
@@ -1,4 +1,4 @@
-package screens;
+package com.bitsavior.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
@@ -10,9 +10,10 @@ import com.bitsavior.game.World;
 public class GameScreen extends ScreenAdapter {
 
 	// private Members
-	BitSavior game;
-	TitleScreen tScreen;
+	private BitSavior game;
 	private World world;
+
+	private int gameLevel = 0;
 
 	
 	/**
@@ -26,13 +27,14 @@ public class GameScreen extends ScreenAdapter {
 	 * @param game game
 	 * @param tScreen to "jump back"
 	 */
-	public GameScreen(BitSavior game, TitleScreen tScreen)
+	public GameScreen(BitSavior game, int level)
 	{
 		this.game = game;
-		if(tScreen==null) System.out.println("tScreen beim erstellen von GameScreen leer");
-		this.tScreen = tScreen;
+		//if(tScreen==null) System.out.println("tScreen beim erstellen von GameScreen leer");
+		//this.tScreen = tScreen;
 		world = new World(GameState.INITIALIZE);    // creating the World
 		world.create();
+		gameLevel = level;
 		if(aTestMode==1) System.out.println("creates new world");
 	}
 
@@ -44,10 +46,9 @@ public class GameScreen extends ScreenAdapter {
 			@Override
 			public boolean keyDown(int keyCode) {
 				if (keyCode == Input.Keys.ESCAPE) {
-					
-					if(tScreen==null) if(aTestMode==1) System.out.println("tScreen leer");
-					game.setScreen(tScreen);
-					
+
+					game.manager.showScreen(Screens.TITLE);
+
 				}
 				return true;
 			}
@@ -63,12 +64,15 @@ public class GameScreen extends ScreenAdapter {
 		{
 			case LOOSE_CAUGHT:
 			case LOOSE_TIMEOUT:
-				game.setScreen(tScreen);
+				game.manager.showScreen(Screens.TITLE);
 				break;
 			default:
 		}
 		world.render(Delta);
 	}
+
+	public int getGameLevel() { return gameLevel; }
+
 
 	@Override
 	public void dispose()
